Use in-content preferences.css for styling.
Use in-content preferences.css to style language manager this makes
language manager look very nice and works great with full themes.

This allows language managers visual experience to grow with each new
version of Firefox.

There should be no conflicts with this as it only affects the content of
language_manager.xul

Will have to see if this passes AMO.